pyproject.toml
[project]
name = "draupnir-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "kyungjunleeme", email = "kyungjunlee.me@gmail.com" }
]
requires-python = ">=3.12.11"
dependencies = [
    "mcp>=1.16.0",
    "streamlit>=1.50.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
draupnir-mcp-server = "draupnir_mcp_server:main"
cnp-ingest = "draupnir_mcp_server.ingest:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/draupnir_mcp_server"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.13.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
